requirements.txt
# python -m pip install --upgrade pip
# pip install -r requirements.txt

onnxruntime>=1.15.0
fastapi>=0.95.0
uvicorn>=0.21.0
pandas>=2.0.0
scikit-learn>=1.2.2
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.24.0
aiohttp>=3.8.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
tensorflow>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=1.4.41
pyjwt>=2.6.0
email-validator>=2.0.0
google-api-python-client
fastapi-mail
httpx
bcrypt
langchain_openai
langchain-community 
langchain-openai 
langgraph
beautifulsoup4
newspaper3k
lxml_html_clean
torch
torchvision
pillow
pydantic>=1.10.0 
typing-extensions>=4.5.0
transformers>=4.49.0
huggingface-hub>=0.26.0,<1.0

# python -m uvicorn app:app --reload --host 0.0.0.0 --port 8000